This commit is controversial.  This message will be long.
I've done a good deal of research in the past several days trying to figure out how to get some of the functionality I know how to do in jQuery into this Angular webpage.  What I've learned is that jQuery and Angular aren't really compatible.  They're technically compatible, but getting them to work together is really a violation of the Angular workflow and long-term causes more work than just learning how to do things via the proper Angular methods.

Great post on that here: https://stackoverflow.com/questions/14994391/thinking-in-angularjs-if-i-have-a-jquery-background

There's not that much jQuery stuff that I want to put in this webpage, but at the same time my current knowledge of Angular is providing literally no immediate benefits because I just don't know enough.  So here's the plan going forward:

 For the present, I'm going to strip the project from Angular and just make a basic webpage with HTML, CSS, and jQuery.  This should allow me to get a website with basic functionality up quickly, as I'd rather have a basic website than no website.  After getting that up, I'll continue to study Angular, with the expectation that I will eventually learn enough that it will make sense to transition the site back into Angular.  Because there's a minimal amount of jQuery I plan to do, I should be able to just jettison the jQuery and start fresh with an Angular architecture.

 That's the plan.

 So, this commit has some minor clean-up stuff in it, but it also has an attempt to get jQuery to work in Angular.  And it does!  Though it has some strange behavior, and I'm assuming this is because what I'm doing really shouldn't be done.  But I wanted to commit this just so it's there.